chore(CleanUpGraph): clean up graph visualization
