+++ packages/react-components/package.json
@@ -67,6 +67,7 @@
     "@types/jest": "^29.4.4",
     "@types/lodash.omitby": "^4.6.7",
     "@types/react": "^18.0.28",
+    "@types/react-cytoscapejs": "^1.2.2",
     "@types/react-dom": "^18.2.4",
     "@types/styled-components": "^5.1.26",
     "@types/uuid": "^9.0.1",
@@ -86,20 +87,21 @@
     "@cloudscape-design/collection-hooks": "^1.0.21",
     "@cloudscape-design/components": "^3.0.277",
     "@cloudscape-design/design-tokens": "^3.0.11",
-    "@iot-app-kit/charts-core": "^1.8.4",
     "@iot-app-kit/charts": "^1.9.0",
+    "@iot-app-kit/charts-core": "^1.8.4",
     "@iot-app-kit/components": "5.7.0",
     "@iot-app-kit/core": "5.7.0",
     "@iot-app-kit/core-util": "5.7.0",
     "@iot-app-kit/source-iottwinmaker": "5.7.0",
     "color": "^4.2.3",
-    "d3-array": "^3.2.3",
     "cytoscape": "3.23.0",
     "cytoscape-cise": "1.0.0",
+    "d3-array": "^3.2.3",
     "d3-shape": "^3.2.0",
     "dompurify": "3.0.3",
     "lodash.omitby": "^4.6.0",
     "parse-duration": "^1.0.3",
+    "react-cytoscapejs": "^2.0.0",
     "styled-components": "^5.3.10",
     "uuid": "^9.0.0",
     "video.js": "7.20.3"
