+++ packages/react-components/src/components/graph/index.ts
@@ -1,0 +1,1 @@
+export { Graph } from './graph';
