+++ packages/react-components/src/components/graph/lib/svgs/icons/MinusIcon/index.tsx
@@ -1,0 +1,11 @@
+import React, { FC, HTMLAttributes } from 'react';
+
+const MinusIcon: FC<HTMLAttributes<SVGElement>> = (props) => {
+  return (
+    <svg viewBox='0 0 100 16' {...props}>
+      <rect width='100' height='16' />
+    </svg>
+  );
+};
+
+export default MinusIcon;
