+++ packages/react-components/src/components/graph/lib/svgs/icons/index.ts
@@ -1,0 +1,4 @@
+export { default as FitIcon } from './FitIcon';
+export { default as MinusIcon } from './MinusIcon';
+export { default as PlusIcon } from './PlusIcon';
+export { default as TargetIcon } from './TargetIcon';
