+++ packages/react-components/src/global.d.ts
@@ -1,0 +1,3 @@
+declare module '*.module.css';
+declare module '*.module.scss';
+declare module 'cytoscape-cise';
