--- packages/react-components/src/graph-visualizer/index.d.ts
@@ -1,3 +1,0 @@
-declare module '*.module.css';
-declare module '*.module.scss';
-declare module 'cytoscape-cise';
