--- packages/react-components/src/graph-visualizer/lib/svgs/icons/index.ts
@@ -1,4 +1,0 @@
-export { FitIcon } from './FitIcon';
-export { MinusIcon } from './MinusIcon';
-export { PlusIcon } from './PlusIcon';
-export { TargetIcon } from './TargetIcon';
