--- packages/react-components/src/graph-visualizer/lib/svgs/icons/styles.module.css
@@ -1,4 +1,0 @@
-.svg {
-  fill: currentcolor;
-  shape-rendering: geometricprecision;
-}
