--- packages/react-components/src/graph-visualizer/styles.module.scss
@@ -1,108 +1,0 @@
-@import './lib/css/colors.css';
-:root {
-  --background-color: transparent;
-  --background-hover-color: var(--color-gray-46);
-  --border-radius: 3px;
-  --border-color: var(--color-gray-54);
-  --border-hover-color: var(--color-gray-54);
-  --menu-z-index: 999;
-  --icon-color: var(--color-gray-26);
-  --icon-hover-color: var(--color-gray-4);
-  --icon-active-color: var(--color-gray-4);
-  --text-color: var(--color-gray-14);
-  --text-hover-color: var(--color-gray-4);
-  --text-active-color: var(--color-gray-4);
-  --transition: background-color var(--transition-duration) ease-out, border-color var(--transition-duration) ease-out,
-    color var(--transition-duration) ease-out, rotate var(--transition-duration) ease-out;
-  font-family: var(--font-family);
-  font-size: 1px;
-  text-rendering: geometricPrecision;
-}
-
-.root {
-  --background-color: var(--color-gray-43);
-  --button-background-color: var(--color-gray-42);
-  --button-background-hover-color: var(--color-gray-38);
-  --button-icon-color: var(--color-gray-16);
-  --button-icon-hover-color: var(--color-gray-4);
-  position: relative;
-  height: 100%;
-  min-height: 1000px;
-  overflow: 'auto';
-}
-
-.content {
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  overflow: 'auto';
-}
-
-.canvas {
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  right: 0;
-  left: 0;
-  z-index: 0;
-}
-
-.controls {
-  position: absolute;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 8rem;
-  width: 40rem;
-  height: 100%;
-  top: 0;
-  right: 20rem;
-  z-index: 1;
-  pointer-events: none;
-}
-
-.button {
-  box-sizing: content-box;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 36rem;
-  height: 36rem;
-  background-color: var(--button-background-color);
-  border-radius: var(--border-radius);
-  box-shadow: 0 0 8px var(--color-gray-50);
-  transition: var(--transition);
-  pointer-events: all;
-}
-
-.button:hover {
-  --button-background-color: var(--button-background-hover-color);
-  --button-icon-color: var(--button-icon-hover-color);
-}
-
-.buttonCenterIcon {
-  width: 20rem;
-  color: var(--button-icon-color);
-  transition: var(--transition);
-}
-
-.buttonFitIcon {
-  width: 18rem;
-  color: var(--button-icon-color);
-  transition: var(--transition);
-}
-
-.buttonZoomInIcon {
-  width: 16rem;
-  color: var(--button-icon-color);
-  transition: var(--transition);
-}
-
-.buttonZoomOutIcon {
-  width: 16rem;
-  color: var(--button-icon-color);
-  transition: var(--transition);
-}
