+++ packages/react-components/tsconfig.build.json
@@ -3,8 +3,4 @@
   "include": [
     "./src"
   ],
-  "exclude": [
-    "**/*.spec.*",
-    "**/*.css"
-  ]
 }
